More outputs and -40 penalty
Penalization for disconnected graphs can be customized as a
linear function of the number of components.

  penalization = -(a + b * n)

Here n is the total number of connected components. The
previous behaviour is obtained with b=10 and a=-10, so
these are the default values.

We try experiments with a flat penalization of -40 and
different hyperparameters for PPO.

Updated the callback to check the rollout performances from
the buffer and plot their distribution.